progress on ensuring Java's long is represented as int64 (for #83)

+++ client_configuration.go
@@ -2,8 +2,8 @@ package ravendb
 
 // ClientConfiguration represents client configuration
 type ClientConfiguration struct {
-	Etag       int  `json:"Etag"`
-	IsDisabled bool `json:"Disabled"`
+	Etag       int64 `json:"Etag"`
+	IsDisabled bool  `json:"Disabled"`
 	// TODO: should this be *int ?
 	MaxNumberOfRequestsPerSession int                 `json:"MaxNumberOfRequestsPerSession"`
 	ReadBalanceBehavior           ReadBalanceBehavior `json:"ReadBalanceBehavior"`
